change the fonction of simulator to accept double value

+++ BL/BlApi/IAdmin.cs
@@ -14,6 +14,6 @@ public interface IAdmin
     void RemoveConfigObserver(Action configObserver);
     void AddClockObserver(Action clockObserver);
     void RemoveClockObserver(Action clockObserver);
-    void StartSimulator(int interval);             // stage 7
+    void StartSimulator(double interval);             // stage 7
     void StopSimulator();                          // stage 7
 }
